feat: Add local patient seeding script and finalize patient view
This commit introduces a new local script for seeding patient data and finalizes the patient view feature for admins.

- A new standalone script `scripts/seed-patients.js` has been created to allow developers to seed the Firestore `patients` collection with 10 sample patients from their local terminal. This replaces the previous approach of seeding patients via a Netlify function.
- The `netlify/functions/seed-users.js` file has been restored to its original state, ensuring that essential user and department data is not affected.
- Instructions on how to use the new seeding script have been added as comments in the file.
- The `ViewPatients.vue` component was also cleaned up by removing a stale comment.
